+++ Tests/TestData.swift
@@ -8,7 +8,7 @@ enum TestData {
     static let currency = Currency.inr
     static let currency0 = Currency.aud
     static let currency1 = Currency.rub
-    static let multiplier: Decimal = 4000.32
+    static let multiplier: Double = 4000.32
     static let multiplier1 = 23423442234.32
     static let rateValue0 = 1.6208
     static let rateValue1 = 1.9612
